rename hasApproveLicense probe
