+++ checks/evaluation/license.go
@@ -20,7 +20,7 @@ import (
 	"github.com/ossf/scorecard/v4/checker"
 	sce "github.com/ossf/scorecard/v4/errors"
 	"github.com/ossf/scorecard/v4/finding"
-	"github.com/ossf/scorecard/v4/probes/hasApprovedLicense"
+	"github.com/ossf/scorecard/v4/probes/hasFSFOrOSIApprovedLicense"
 	"github.com/ossf/scorecard/v4/probes/hasLicenseFile"
 	"github.com/ossf/scorecard/v4/probes/hasLicenseFileAtTopDir"
 )
@@ -33,7 +33,7 @@ func License(name string,
 	// We have 3 unique probes, each should have a finding.
 	expectedProbes := []string{
 		hasLicenseFile.Probe,
-		hasApprovedLicense.Probe,
+		hasFSFOrOSIApprovedLicense.Probe,
 		hasLicenseFileAtTopDir.Probe,
 	}
 
@@ -49,7 +49,7 @@ func License(name string,
 		f := &findings[i]
 		if f.Outcome == finding.OutcomePositive {
 			switch f.Probe {
-			case hasApprovedLicense.Probe:
+			case hasFSFOrOSIApprovedLicense.Probe:
 				dl.Info(&checker.LogMessage{
 					Type:   finding.FileTypeSource,
 					Offset: 1,
@@ -82,7 +82,7 @@ func License(name string,
 						Text:   "License file found in unexpected location",
 					})
 				}
-			case hasApprovedLicense.Probe:
+			case hasFSFOrOSIApprovedLicense.Probe:
 				dl.Warn(&checker.LogMessage{
 					Type:   finding.FileTypeSource,
 					Offset: 1,
